Remove translator() method
